Split single-atom handling into three.
